+++ Utilities/ScriptingLibrary/ScriptingEngine.cs
@@ -27,7 +27,8 @@ namespace Sheepy.PhoenixPt.ScriptingLibrary {
                HostTypes.Add( "Log", typeof( LogHelper ) );
                HostTypes.Add( "Repo", typeof( RepoHelper ) );
                HostTypes.Add( "Patch", typeof( PatchHelper ) );
-               HostTypes.Add( "Reflection", typeof( ReflectionHelper ) );
+               HostTypes.Add( "Extensions", typeof( ExtensionHelper ) );
+               HostTypes.Add( "Espy", typeof( Espy ) );
                HostTypes.Add( "console", typeof( ConsoleHelper ) );
                //HostTypes.Add( "Patch", typeof( PatchHelper ) );
                HostTypes.Add( "Damage", typeof( DamageHelpers ) );